split default mapping tests
